# openapi.yaml
openapi: "3.1.0"
info:
  title: kube-phoenix API
  description: |
    REST API for the kube-phoenix cluster sleep/wake scheduler.

    All `/api/*` and `/ws/*` endpoints require session-based authentication
    via HTTP-only cookies. `/healthz` and `/metrics` are always open.
    `/api/auth/login` and `/api/auth/oidc/*` are unauthenticated.

    Authenticate by calling `POST /api/auth/login` with username/password.
    The server sets a `__kp_session` HTTP-only cookie and a `__kp_csrf`
    cookie. Include the CSRF token as `X-CSRF-Token` header on mutations.

    WebSocket connections use cookies automatically on same-origin upgrades.
  version: "1.1.0"
  contact:
    name: kube-phoenix
    url: https://github.com/MacXsimilian/kube-phoenix
  license:
    name: Apache 2.0
    url: https://github.com/MacXsimilian/kube-phoenix/blob/master/LICENSE
...
servers:
  - url: /
    description: Same-origin (served by the Go binary)
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __kp_session
  schemas:
...
    ScheduleCreate:
      description: Request body for creating a new schedule.
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [scale_down, scale_up]
        cronExpr:
          type: string
        timezone:
          type: string
          default: UTC
        mode:
          type: string
          enum: [plan, apply]
          default: plan
        enabled:
          type: boolean
          default: false
        namespaceFilter:
          type: string
          default: ""
        timeoutMinutes:
          type: integer
          default: 0
      required:
        - name
        - type
        - cronExpr
...
    Error:
      description: Standard error response.
      type: object
      properties:
        error:
          type: string

    # ── Policy model ────────────────────────────────────────────────────────
...
paths:
...
  /api/schedules:
    get:
      operationId: listSchedules
      summary: List schedules
      description: Returns all schedules ordered by position. Each schedule includes a computed `nextRun` field.
      tags: [Schedules]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createSchedule
      summary: Create schedule
      description: Creates a new schedule. Defaults to disabled, plan mode, UTC timezone.
      tags: [Schedules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/schedules/reorder:
    put:
      operationId: reorderSchedules
      summary: Reorder schedules
      description: |
        Bulk-updates the `position` field for schedules of a given type.
        Only IDs belonging to the specified type are affected — mismatched IDs are silently ignored.
        Returns the full updated schedule list with `nextRun` fields.
      tags: [Schedules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [scale_down, scale_up]
                ids:
                  type: array
                  items:
                    type: integer
                  example: [3, 1, 2]
              required:
                - type
                - ids
      responses:
        "200":
          description: Reordered — returns updated schedule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/schedules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getSchedule
      summary: Get schedule
      description: Returns a single schedule by ID, including the computed `nextRun` field.
      tags: [Schedules]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Invalid schedule ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateSchedule
      summary: Update schedule
      description: Partial update — only provided fields are written. `type` is immutable and ignored.
      tags: [Schedules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSchedule
      summary: Delete schedule
      description: Deletes a schedule by ID and reloads the scheduler.
      tags: [Schedules]
      responses:
        "204":
          description: Deleted
        "400":
          description: Invalid schedule ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/executions:
    get:
      operationId: listExecutions
      summary: List executions
      description: Returns a paginated list of executions, ordered by start time descending.
      tags: [Executions]
      parameters:
        - name: schedule_id
          in: query
          schema:
            type: integer
          description: Filter by schedule ID.
        - name: status
          in: query
          schema:
            type: string
            enum: [running, success, failed]
          description: Filter by execution status.
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: Zero-based page index.
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Number of items per page (max 100).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedExecutions"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/executions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getExecution
      summary: Get execution
      description: Returns a single execution by ID with the associated schedule preloaded.
      tags: [Executions]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "400":
          description: Invalid execution ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/executions/{id}/logs:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getExecutionLogs
      summary: Get execution logs
      description: Returns all log lines for an execution ordered by sequence number. For live streaming during a running execution, use the WebSocket endpoint instead.
      tags: [Executions]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogLine"
        "400":
          description: Invalid execution ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/overview:
    get:
      operationId: getOverview
      summary: Dashboard overview
      description: |
        Pre-aggregated cluster summary for the Overview page.
        Reads from the in-memory ClusterCache — no Kubernetes or database I/O on the hot path.
      tags: [Overview]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewResponse"
        "401":
          description: Unauthorized
  /api/cluster/stream:
    get:
      operationId: streamCluster
      summary: SSE overview stream
      description: |
        Server-Sent Events stream. Sends the current overview immediately on connect,
        then pushes a new `data:` event on every ClusterCache refresh (~10 s cadence).
        Each event payload is a JSON-serialised `OverviewResponse`.
      tags: [Overview]
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
                description: Each `data:` line contains a JSON OverviewResponse.
        "401":
          description: Unauthorized
        "503":
          description: Cluster cache unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cluster/workloads:
    get:
      operationId: listWorkloads
      summary: List workloads
      description: Returns all Deployments and StatefulSets with sleep status derived from the `previous-replicas` annotation.
      tags: [Cluster]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkloadResponse"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Kubernetes client unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cluster/nodes:
    get:
      operationId: listNodes
      summary: List nodes
      description: Returns all nodes enriched with non-DaemonSet pod counts, resource requests/allocatable, and guardrail protection status.
      tags: [Cluster]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NodeResponse"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Kubernetes client unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cluster/nodes/{name}/pods:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Kubernetes node name.
    get:
      operationId: listNodePods
      summary: List pods on a node
      description: Returns non-DaemonSet pods on the specified node. Resolves ReplicaSet owners to top-level Deployment/StatefulSet names.
      tags: [Cluster]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NodePodResponse"
        "400":
          description: Invalid node name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Kubernetes client unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cluster/pods/{namespace}/{name}:
    parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPodDetail
      summary: Get pod detail
      description: |
        Returns comprehensive pod information: containers with resource requests/limits/usage,
        conditions, recent Kubernetes events, and live CPU/memory from the Metrics Server
        (degrades gracefully if Metrics Server is absent).
      tags: [Cluster]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PodDetailResponse"
        "400":
          description: Invalid resource name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Kubernetes client unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cluster/pods/{namespace}/{name}/logs:
    parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPodLogs
      summary: Stream container logs
      description: |
        Proxies the Kubernetes pod logs API. Returns plain-text log output.
        When `follow=true`, the response uses chunked transfer encoding and
        streams new log lines as the container writes them. The connection
        stays open until the client disconnects or the container terminates.
        No database involvement — logs are streamed directly from the K8s API.
      tags: [Cluster]
      parameters:
        - name: container
          in: query
          description: Container name (defaults to first container)
          schema:
            type: string
        - name: tailLines
          in: query
          description: Number of lines from the end to start from (default 500, max 10000)
          schema:
            type: integer
            default: 500
        - name: follow
          in: query
          description: Stream live logs (chunked transfer encoding)
          schema:
            type: boolean
            default: false
        - name: previous
          in: query
          description: Return logs from the previously terminated container instance
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Plain-text log output (chunked when follow=true)
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid resource name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Kubernetes client unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cluster/workloads/{namespace}/{kind}/{name}/pods:
    parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: kind
        in: path
        required: true
        schema:
          type: string
          enum: [Deployment, StatefulSet]
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listWorkloadPods
      summary: List pods for a workload
      description: Returns non-DaemonSet pods owned by the specified Deployment or StatefulSet.
      tags: [Cluster]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NodePodResponse"
        "400":
          description: Invalid resource name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Kubernetes client unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/guardrails:
    get:
      operationId: getGuardrails
      summary: Get guardrails
      description: Returns the current guardrail configuration.
      tags: [Guardrails]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guardrails"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateGuardrails
      summary: Update guardrails
      description: Partial update — only provided fields are written.
      tags: [Guardrails]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardrailsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guardrails"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/trigger:
    post:
      operationId: triggerSchedule
      summary: Manually trigger a schedule
      description: Runs the specified schedule immediately, outside its cron cadence. Returns the new execution ID.
      tags: [Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduleId:
                  type: integer
                mode:
                  type: string
                  enum: [plan, apply]
                  description: Overrides the schedule's configured mode for this run.
              required:
                - scheduleId
                - mode
      responses:
        "202":
          description: Accepted — execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Schedule not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/reset-db:
    post:
      operationId: resetDatabase
      summary: Reset database
      description: |
        Drops all tables and reseeds defaults. Streams progress as newline-delimited JSON (NDJSON).
        Requires confirmation in the request body.

        **This operation is irreversible. All schedules, executions, and log history will be deleted.**
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                confirm:
                  type: string
                  enum: ["RESET DATABASE"]
              required:
                - confirm
      responses:
        "200":
          description: Streaming NDJSON progress
          content:
            application/x-ndjson:
              schema:
                $ref: "#/components/schemas/ResetEvent"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "422":
          description: Missing or incorrect confirmation string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...